+++ atlas/foundations_internal/src/foundations_internal/foundations_context.py
@@ -1,6 +1,35 @@
-from .pipeline import Pipeline
+from foundations_internal.provenance import Provenance
 
 class FoundationsContext(object):
+    class Pipeline(object):
+        class PipelineContext(object):
+
+            def __init__(self):
+                self._file_name = None
+                self.provenance = Provenance()
+
+            @property
+            def file_name(self):
+                if not self._file_name:
+                    raise ValueError('Job ID is currently undefined, please set before retrieving')
+                return self._file_name
+
+            @file_name.setter
+            def file_name(self, value):
+                self._file_name = value
+
+            @property
+            def job_id(self):
+                return self.file_name
+
+        def __init__(self, pipeline_context=None):
+            if pipeline_context is None:
+                pipeline_context = FoundationsContext.Pipeline.PipelineContext()
+            self._pipeline_context = pipeline_context
+
+        def pipeline_context(self):
+            return self._pipeline_context
+
     """The global state for all staging related functionality for Foundations.
     This is where everything awesome begins!!!
 
@@ -10,7 +39,7 @@ class FoundationsContext(object):
 
     def __init__(self, pipeline=None):
         if pipeline is None:
-            pipeline = Pipeline()
+            pipeline = FoundationsContext.Pipeline()
         self._pipeline = pipeline
         self._job_resources = self._default_job_resources()
